AI_SERVICE/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
transformers==4.39.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.9
numpy==1.26.4
